+++ src/constructors/TwistStamped.h
@@ -3,8 +3,8 @@
 #include <geometry_msgs/Vector3.h>
 #include <std_msgs/Header.h>
 
-#include "helpers/Header.h"
-#include "helpers/Twist.h"
+#include "constructors/Header.h"
+#include "constructors/Twist.h"
 
 inline geometry_msgs::TwistStamped TwistStamped(std_msgs::Header header,
                                                 geometry_msgs::Twist twist) {
